add (test): setup regression tests on real hardware
