+++ .pre-commit-config.yaml
@@ -15,5 +15,5 @@ repos:
     hooks:
 #    -   id: clang-format
 #        args: ["-style=file"]
-    -   id: cpplint
+#    -   id: cpplint
     -   id: cppcheck
