+++ platformio_ci.ini
@@ -7,7 +7,7 @@ lib_deps =
     ArduinoFake=https://github.com/FabioBatSilva/ArduinoFake
 build_flags =
     -DPIO_NATIVE_TESTING
-    -std=gnu++17
+#    -std=gnu++17
     -frtti
     -fexceptions
 #build_unflags =
